Set start date on first episode
We set the start date when we watch the first episode.  Doing it on the
transition away from 'plan to watch' skips the case when adding a show
as watching.  I don't think there's a case where watching the first
episode shouldn't set the start date, and the same for the inverse.