watch: Pick file instead of using version.
This is a huge commit/refactor of the watch command.

The driving change is that instead of guessing the file to play using
the version, instead prompt the user to choose which file to play.

This is more flexible, and it ends up making the code a lot cleaner,
hence the fat commit and refactor.